+++ src/android/LocationProviderEnum.java
@@ -10,16 +10,16 @@ This is a new class
 package com.marianhello.cordova.bgloc;
 
 /**
- * ServiceProviderEnum
+ * LocationProviderEnum
  */
-public enum ServiceProviderEnum
+public enum LocationProviderEnum
 {
     ANDROID_DISTANCE_FILTER(0),
     ANDROID_FUSED_LOCATION(1);
 
     public final int id;
 
-    private ServiceProviderEnum(int id) {
+    private LocationProviderEnum(int id) {
         this.id = id;
     }
 
@@ -27,8 +27,8 @@ public enum ServiceProviderEnum
         return this.id;
     }
 
-    public static ServiceProviderEnum forInt (int id) {
-        for (ServiceProviderEnum provider : values()) {
+    public static LocationProviderEnum forInt (int id) {
+        for (LocationProviderEnum provider : values()) {
             if (provider.id == id) {
                 return provider;
             }