+++ src/android/data/sqlite/ConfigurationContract.java
@@ -23,7 +23,7 @@ public final class ConfigurationContract {
         public static final String COLUMN_NAME_STOP_TERMINATE = "stop_terminate";
         public static final String COLUMN_NAME_START_BOOT = "start_boot";
         public static final String COLUMN_NAME_START_FOREGROUND = "start_foreground";
-        public static final String COLUMN_NAME_SERVICE_PROVIDER = "service_provider";
+        public static final String COLUMN_NAME_LOCATION_PROVIDER = "service_provider";
         public static final String COLUMN_NAME_INTERVAL = "interval";
         public static final String COLUMN_NAME_FASTEST_INTERVAL = "fastest_interval";
         public static final String COLUMN_NAME_ACTIVITIES_INTERVAL = "activities_interval";